requirements.txt
Flask>=2.0
pandas>=1.3
scikit-learn>=1.0
gunicorn>=20.0
pytest>=7.0
joblib>=1.0
numpy>=1.21
python-dotenv>=0.21
duckduckgo-search>=2.6.1
google-generative-ai>=0.3.0